build(tools): 20 个工具全量纳入 all/tools/check-tools + CI 门禁；三编译器告警清零
问题：tools/ 层长期「零自动编译」——23 个 .c 里 11 个没有任何二进制规则、
9 个虽有规则却不在 all: 里；make linux（CI 用的就是它）只编 4 个二进制。

- Makefile：新增 TOOL_BINS（20 个工具二进制，all 与 tools 同源）+ tools: 聚合目标
  + check-tools: 门禁（SSOT 清单断言全产出，禁 CI 手抄）；补齐 11 条缺失的二进制
  规则与 11 个 phony 别名；all: 并入 $(TOOL_BINS)。
- batch_learn_lowmem 规则修正：旧规则把 -DLOW_MEM 写在链接行（对已编译 .o 无效），
  改为编译自己的翻译单元。
- Makefile 注释 14 处 U+FFFD 乱码复原（纯注释）。
- ci.yml：build-x86_64 / build-arm 各加一步 make check-tools。

真 bug（三编译器交叉扫出，详见 changelogs/075）：
- quick_chat.c:35 缺 (struct ExternalDict*) 强转 —— GCC14+ 直接 error。
- reader.c is_sent_end()：char 在 x86_64 有符号 ⇒ p[1]==0xBC 恒假，中文句末标点
  判定在 x86 上完全失效。改走 const unsigned char*。
- .bak 缓冲截断 3 处（batch_learn / build_cross_links / hebbian_pretrain）：
  char bak[520] + snprintf(bak,519,"%s.bak",path) 而 path 上限 4096。
- merge_state.c：14 处 fread 未检查 → 新增 rd_or_die()（截断即报错退出）；
  17 处 int→uint32_t + 4 处 %d→%u。
- 其余：hebbian_pretrain fread/ftell/malloc 守卫；digital_life system()（(void) 不能
  抑制 -Wunused-result，须赋值）；若干 -Wall -Wextra 告警。

验证：WSL gcc15.2 / Pi gcc14.2 / armbian gcc13.3 三处 make clean && make all
全量构建 0 error 0 warning。

+++ tools/seed_teacher.c
@@ -127,7 +127,7 @@ int main(int argc, char* argv[]) {
     // 批量注入对话模式
     int new_count = 0;
     int skip_count = 0;
-    for (int i = 0; i < PATTERN_COUNT; i++) {
+    for (size_t i = 0; i < PATTERN_COUNT; i++) {
         char key[256];
         snprintf(key, sizeof(key), "response:%s", patterns[i][0]);
